Week 6/Task 1/requirements.txt
# AWS and Bedrock
boto3>=1.34.0
botocore>=1.34.0

# LangChain
langchain>=0.1.0
langchain-aws>=0.1.0
langchain-community>=0.0.20

# NeMo Guardrails
nemoguardrails>=0.8.0
nest-asyncio>=1.5.0

# Observability
langfuse>=2.0.0

# Evaluation
agentevals>=0.1.0
sentence-transformers>=2.0.0

# Core dependencies
typing-extensions>=4.0.0

# Optional but recommended
python-dotenv>=1.0.0
